+++ Backend/src/main/java/com/Project/Bookstore/Service/VoucherService.java
@@ -18,13 +18,33 @@ public class VoucherService {
     public List<Voucher> getAllVouchers() {
         return this.voucherRepository.findAll();
     }
-    public Voucher getVoucherById(Integer id) {
-        return this.voucherRepository.findById(id).orElse(null);
+    public Voucher getVoucherById(String id) {
+        return this.voucherRepository.findById(id).
+                orElseThrow(()-> new RuntimeException("Voucher not found!"));
     }
     public Voucher saveVoucher(Voucher voucher) {
+        if(voucher.getPercent() <0 || voucher.getPercent() > 100){
+            throw new RuntimeException("Percentage must be between 0 and 100");
+        }
+        if(voucher.getQuantity() <0){
+            throw new RuntimeException("Quantity must be greater than 0");
+        }
         return this.voucherRepository.save(voucher);
     }
-    public void deleteVoucher(Integer id) {
+    public Voucher updateVoucher(Voucher voucher) {
+        if(!this.voucherRepository.existsById(voucher.getCode()))
+            throw new RuntimeException("Voucher not found!");
+        if(voucher.getPercent() <0 || voucher.getPercent() > 100){
+            throw new RuntimeException("Percentage must be between 0 and 100");
+        }
+        if(voucher.getQuantity() <0){
+            throw new RuntimeException("Quantity must be greater than 0");
+        }
+        return this.voucherRepository.save(voucher);
+    }
+    public void deleteVoucher(String id) {
+        if(!this.voucherRepository.existsById(id))
+            throw new RuntimeException("Voucher not found!");
         this.voucherRepository.deleteById(id);
     }
 }
